Improve the performance of simple dates, take 1
Simples dates are common in many cases. Performance analysis showed that this can be a performance hotspot during harvesting.
This is due to the design of UUIDMapper, which loads all metadata for a harvester for every new batch. This can not be easily changed, but we can improve the performance here, which also helps other code paths using this method.
This change improves the performance of the method by a factor of about 1.3x